+++ CHANGELOG.md
@@ -7,7 +7,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 ## Firmware
 
 ### [Unreleased]
-- Bitcoin: warn if the transaction fee is higher than 10% of the coins sent
+- Bitcoin, Ethereum: warn if the transaction fee is higher than 10% of the coins sent
 - ETH Testnets: add Goerli and remove deprecated Rinkeby and Ropsten
 
 ### 9.13.1